:recycle: Re-structure fileupload project
