+++ @navikt/core/react/src/form/file-upload/helpers/download-file.ts
