+++ @navikt/core/react/src/form/file-upload/helpers/format-file-size.test.ts
