+++ @navikt/core/react/src/form/file-upload/helpers/is-accepted-file-type.ts
