+++ @navikt/core/react/src/form/file-upload/helpers/is-accepted-size.test.ts
