+++ @navikt/core/react/src/form/file-upload/helpers/validate-files.ts
