+++ @navikt/core/react/src/form/file-upload/item-root/FileUploadItemRoot.types.ts
