+++ @navikt/core/react/src/form/file-upload/root/FileUploadRoot.context.tsx
@@ -1,5 +1,5 @@
-import { createStrictContext } from "../../utils/helpers";
-import { ComponentTranslation } from "../../utils/i18n/i18n.types";
+import { createStrictContext } from "../../../utils/helpers";
+import { ComponentTranslation } from "../../../utils/i18n/i18n.types";
 
 export type FileUploadLocaleContextProps = {
   translations?: ComponentTranslation<"FileUpload">;
